modified

+++ scripts/plugins/gulp-s3-sync.ts
@@ -5,10 +5,15 @@ import * as xmlNodes from "xml-nodes";
 import * as xmlObjects from "xml-objects";
 import * as pumpify from "pumpify";
 import * as through from "through2";
-
 import * as AWS from "aws-sdk";
+import Utils from "../utils";
 
-function buildDeleteMultiple(bucket, keys) {
+const _DEFAULT_OPTION = {
+  region: "ap-northeast-1",
+  profile: "default"
+};
+
+function buildDeleteParameter(bucketName, keys) {
   if (!keys || !keys.length) return;
 
   var deleteObjects = keys.map(function(k) {
@@ -16,7 +21,7 @@ function buildDeleteMultiple(bucket, keys) {
   });
 
   return {
-    Bucket: bucket,
+    Bucket: bucketName,
     Delete: {
       Objects: deleteObjects
     }
@@ -37,30 +42,42 @@ class S3sync {
 
   private client: AWS.S3;
 
-  constructor(options?) {
+  constructor(options = _DEFAULT_OPTION) {
     this.client = new AWS.S3({
       region: options.region,
       credentials: new AWS.SharedIniFileCredentials({ profile: options.profile })
     });
   }
 
-  sync(bucket, prefix = "") {
+  sync(bucketName: string, prefix: string = "") {
     const s = new Stream.Transform({ objectMode: true })
     let currentFiles = {};
     currentFiles[ prefix + "/" ] = true;
 
     s._transform = (file, encoding, cb) => {
       currentFiles[ path.join(prefix, path.basename(file.path)) ] = true;
-      Object.assign(file, { s3: { path: path.join(prefix, path.basename(file.path)) ,state: "uploaded", etag: "dummy" } });
+      Object.assign(file, { s3: { path: path.join(prefix, path.basename(file.path)) ,state: "uploaded", etag: Utils.genETag(file.contents) } });
       s.push(file);
-      this.client.putObject({
-        Bucket: bucket,
-        Key: path.join(prefix, path.basename(file.path)),
-        Body: file.contents
-      }).promise().then(result => {
-        // console.log(result);
+      this.client.headObject({ Bucket: bucketName, Key: file.s3.path }).promise()
+      .then(result => {
+        console.log("s3: %s | local: %s", result.ETag, file.s3.etag);
+        if (result.ETag.replace(/[\"]/g,"") != file.s3.etag) {
+          this.client.putObject({
+            Bucket: bucketName,
+            Key: path.join(prefix, path.basename(file.path)),
+            Body: file.contents
+          }).promise().then(result => {
+            console.log("Uploaded: %s", result.ETag);
+            cb();
+          });
+        } else {
+          console.log("No Changed: %s", result.ETag);
+          cb();
+        }
+      }).catch(error => {
+        console.error(error)
+        if (error && [403, 404].indexOf(error.statusCode) < 0) return cb(error);
       });
-      cb();
     };
 
     s._flush = (cb) => {
@@ -68,7 +85,7 @@ class S3sync {
       let toPut= [];
       let toDelete = [];
       const lister = c
-        .listObjects({ Bucket: bucket, Prefix: prefix })
+        .listObjects({ Bucket: bucketName, Prefix: prefix })
         .createReadStream()
         .pipe(converter("Key"));
 
@@ -87,7 +104,7 @@ class S3sync {
 
       lister.on('end', function() {
         if (!toDelete.length) return cb();
-        c.deleteObjects(buildDeleteMultiple(bucket, toDelete), cb);
+        c.deleteObjects(buildDeleteParameter(bucketName, toDelete), cb);
       });
     };
